refactor(tools): separate write-surface discovery from registry checking
Adding the intent-compiler criterion pushed the checker past the 650-line
ceiling for tool sources. The file had two jobs anyway: read the repository to
find every write surface that exists, and compare that evidence against the
registry. Move the first into its own module so each side can be read on its
own terms.
